hvf_trader/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
sqlalchemy>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pytest>=7.4.0
